+++ failsafe/__init__.py
@@ -7,4 +7,4 @@ import logging
 
 logging.getLogger(__name__).addHandler(logging.NullHandler())
 
-__version__ = '0.1.3'
+__version__ = '0.2.0'